+++ NEWS.md
@@ -12,13 +12,21 @@
 
 ## New funs
 
-* Added `add_shinyproxy_dockerfile()` to create a Dockerfile from a DESCRIPTION, to be used in Shiny Proxy.
+* Added `add_dockerfile()` to create a Dockerfile from a DESCRIPTION.
+
+* Added `add_dockerfile_shinyproxy()` to create a Dockerfile from a DESCRIPTION, to be used in Shiny Proxy.
+
+* Added `add_dockerfile_heroku()` to create a Dockerfile from a DESCRIPTION, to be used with Heroku.
 
 * `add_css_file()`, `add_js_file()` and `add_js_handler()` create a CSS, JS, and JS with Shiny custom handler files.
 
 ## Removed
 
 * `use_utils_prod` is now included in golem so you don't have to explicitely include them
+
+## Docs 
+
+* Golem now has four vignettes
 
 # golem 0.0.1.0002